+++ packages/jaspr_cli/lib/src/commands/clean_command.dart
@@ -1,5 +1,7 @@
 import 'dart:io';
 
+import '../process_runner.dart';
+import '../project.dart';
 import 'base_command.dart';
 
 class CleanCommand extends BaseCommand {
@@ -39,14 +41,14 @@ class CleanCommand extends BaseCommand {
       }
 
       logger.write("Running 'dart run build_runner clean'...");
-      await Process.run('dart', ['run', 'build_runner', 'clean']);
+      await ProcessRunner.instance.run(dartExecutable, ['run', 'build_runner', 'clean']);
     }
 
     // TODO support windows
     if (Platform.isMacOS || Platform.isLinux) {
       final pids = await findRunawayProcesses([
-        // server, vm, build_runner, flutter
-        '8080', '8181', '5467', '5678',
+        // server, vm, proxy, flutter
+        defaultServePort, '8181', serverProxyPort, flutterProxyPort,
       ]);
 
       if (pids.isNotEmpty) {
@@ -76,7 +78,7 @@ Future<List<int>> findRunawayProcesses(List<String> ports) async {
   final pids = <int>[];
 
   for (final port in ports) {
-    final proc = await Process.run('lsof', ['-i', ':$port']);
+    final proc = await ProcessRunner.instance.run('lsof', ['-i', ':$port']);
 
     if (proc.exitCode == 0) {
       pids.addAll(