+++ packages/jaspr_cli/lib/src/domains/scopes_domain.dart
@@ -16,8 +16,10 @@ import 'package:path/path.dart' as path;
 import 'package:watcher/watcher.dart';
 import 'package:yaml/yaml.dart';
 
-import '../helpers/daemon_helper.dart';
+import '../daemon/daemon.dart';
+import '../daemon/domain.dart';
 import '../logging.dart';
+import '../project.dart';
 
 class ScopesDomain extends Domain {
   ScopesDomain(Daemon daemon, this.logger) : super(daemon, 'scopes') {
@@ -85,6 +87,7 @@ class ScopesDomain extends Domain {
     _collection = AnalysisContextCollection(
       includedPaths: entryPaths,
       resourceProvider: resourceProvider ?? PhysicalResourceProvider.INSTANCE,
+      sdkPath: dartSdkDir,
     );
 
     if (resourceProvider == null) {