+++ packages/jaspr_cli/lib/src/project.dart
@@ -1,12 +1,13 @@
-import 'dart:io' as io;
+import 'dart:io';
 
-import 'package:file/file.dart';
+import 'package:file/file.dart' hide FileSystemEntity;
 import 'package:file/local.dart';
 import 'package:path/path.dart' as path;
 import 'package:pub_semver/pub_semver.dart';
 import 'package:yaml/yaml.dart';
 
 import 'logging.dart';
+import 'process_runner.dart';
 import 'version.dart';
 
 enum JasprMode { static, server, client }
@@ -20,7 +21,7 @@ extension JasprModeExtension on JasprMode {
 class Project {
   Project(this.logger, {FileSystem? fs, Never Function(int)? exitFn})
     : _fs = fs ?? const LocalFileSystem(),
-      _exitFn = exitFn ?? io.exit;
+      _exitFn = exitFn ?? exit;
 
   final Logger logger;
   final FileSystem _fs;
@@ -96,7 +97,7 @@ class Project {
       // ok
     } else {
       final log = logger.logger;
-      if (log == null) {
+      if (log == null || !stdout.hasTerminal) {
         logger.write(
           'Missing dependency on jaspr_builder in pubspec.yaml file. Make sure to add jaspr_builder to your dev_dependencies.',
           tag: Tag.cli,
@@ -108,7 +109,7 @@ class Project {
           defaultValue: true,
         );
         if (result) {
-          final result = io.Process.runSync('dart', ['pub', 'add', '--dev', 'jaspr_builder']);
+          final result = ProcessRunner.instance.runSync(dartExecutable, ['pub', 'add', '--dev', 'jaspr_builder']);
           if (result.exitCode != 0) {
             log.err(result.stderr as String?);
             logger.write(
@@ -291,3 +292,66 @@ class Project {
     }
   }
 }
+
+const defaultServePort = '8080';
+const serverProxyPort = '5567';
+const flutterProxyPort = '5678';
+
+// The path to the Dart executable in either the Dart or Flutter SDK.
+final dartExecutable = () {
+  final String? executable;
+  if (Platform.isWindows) {
+    // Use 'where.exe' to support powershell as well
+    final result = (ProcessRunner.instance.runSync('where.exe', ['dart.bat', 'dart.exe'])).stdout.toString();
+    executable = result.split(RegExp('(\r\n|\r|\n)')).where((s) => !s.contains('Could not find')).firstOrNull?.trim();
+  } else {
+    executable = (ProcessRunner.instance.runSync('which', ['dart'])).stdout.toString().trim();
+  }
+
+  if (executable == null || executable.isEmpty) {
+    throw Exception('Could not find Dart executable. Make sure Dart is installed and added to your PATH.');
+  }
+
+  bool isSdkExecutable(String executable) {
+    final maybeSdkDir = path.dirname(path.dirname(executable));
+    return FileSystemEntity.isFileSync(path.join(maybeSdkDir, 'version')) &&
+        path.basename(path.dirname(executable)) == 'bin';
+  }
+
+  if (isSdkExecutable(executable)) {
+    return executable;
+  }
+
+  final maybeFlutterDartExecutable = path.join(
+    path.dirname(executable),
+    'cache',
+    'dart-sdk',
+    'bin',
+    Platform.isWindows ? 'dart.exe' : 'dart',
+  );
+  if (isSdkExecutable(maybeFlutterDartExecutable)) {
+    return maybeFlutterDartExecutable;
+  }
+
+  throw Exception(
+    'Found Dart executable at "$executable", but failed to verify the surrounding Dart SDK.\n'
+    'Make sure "${Platform.isWindows ? 'where.exe dart.bat dart.exe' : 'which dart'}" resolves to the Dart executable inside the Dart or Flutter SDK directory.',
+  );
+}();
+
+/// The path to the root directory of the SDK.
+final String dartSdkDir = path.dirname(path.dirname(dartExecutable));
+
+final dartSdkVersion = () {
+  final result = ProcessRunner.instance.runSync(dartExecutable, ['--version']);
+  if (result.exitCode != 0) {
+    return 'unknown';
+  }
+  final output = result.stdout.toString().trim();
+  if (output.startsWith('Dart SDK version:')) {
+    return output.substring('Dart SDK version:'.length).trim();
+  }
+  return 'unknown';
+}();
+
+final String dartDevToolsPath = path.join(dartSdkDir, 'bin', 'resources', 'devtools');