+++ packages/jaspr_cli/test/integration/commands_integration_test.dart
@@ -21,7 +21,7 @@ void main() {
     final runner = setupRunner();
 
     for (final variant in allVariants) {
-      test(variant.name, tags: ['cli', variant.tag], () async {
+      test(variant.name, tags: ['cli-integration-test', variant.tag], () async {
         await runner.run('create --no-pub-get -v ${variant.options} myapp', dir: dirs.root);
 
         for (final f in variant.files) {